Add sphinx-design cards to Usage landing page
This PR does some basic reorganization of some items in the Usage
section of the docs, to bring more visibility to the Layers how-to
guides and the napari workshop materials.

This also adds a mention of the new napari-workshop-template repository.

Included in this change is the move from sphinx-tabs/sphinx-panels to
sphinx-design, as the first two packages are being deprecated in favor
of the latter.