Delete vestigial docs/requirements.txt (#270)
# Description
This PR removes the 2nd, nested `requirements.txt` in `docs/docs`, which
is out of sync with the main one and as far as I can tell isn't used in
any workflow. It's probably from when that `docs` folder was in napari
repo.

**Maybe** it has some use in the cherry-picking napari releases? But I
don't think so based on my look at the v0.4.19 branch.