Update headless-gui action in build (fix screenshots) (#294)
# References and relevant issues
~~Hopefully f i x e s~~ Pretty sure now this fixes both #283 and #285 

# Description
This configures the
[headless-gui](https://github.com/aganders3/headless-gui) action to not
run the tiling window manager it uses by default. My thinking is the
tiling behavior in the window manager is causing issues with certain
screenshots. See
https://github.com/napari/docs/issues/283#issuecomment-1841621067 for
more details.

This also updates the action to a new version with a more common default
screen size. This change in bit-depth fixes the messed up magicgui
screenshot.

Edit: I'm still not _totally_ clear why the stable docs would look okay
but I think this is worth trying.

---------

Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>