Add section on extending napari (#286)
# References and relevant issues
closes #244
closes #239

# Description
This was originally meant to be a small PR closing #244, but has been
extended to adding an 'extending napari' section (as it was difficult to
break it up to smaller PRs), loosely based off of @DragaDoncila slides
from
[here](https://github.com/napari/docs/issues/244#issuecomment-1831360814).

* Add 'Extending napari' section under 'How-to guides' - covers widgets,
~intro to event loop~ and how to connect your how events
* "Hooking up your own events" and ~"An introduction to the event loop
in napari"~ pages mostly unchanged, just moved (edit: 'An introduction
to the event loop in napari' will remain where it is, see:
https://github.com/napari/docs/pull/286#discussion_r1418490112)
* Big changes to the "Using magicgui in napari" page:
* This is now documents how to build widgets with magicgui and by
creating a widget class. Also describes how to add widget to the viewer
and how to adapt the widget to be a plugin widget contribution. Adds
more extensive intro.
* Adds lots of cross references to magicgui, the plugins section and the
cookie-cutter

---------

Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>