Fix broken links and references (#338)
# References and relevant issues
Closes #331

# Description
Fixes a few references that were creating "duplicate target" warnings
when building the docs. Also fixes links in team page and a few other
pages.

A few references are renamed, because they trigger duplicate reference
warnings. For example, if a file is called "benchmarks.md" and its
reference targe is `(benchmarks)=`, the link would have to be marked as
either the `ref` role or the `doc` role. To make it easier to link, I've
changed the reference name to `napari-benchmarks`. There's a couple
other instances of the same issue.

This is meant to guarantee the links will still be valid even after
reorganizations such as the recent contributing docs reorg.