Using Dask and Napari to view and process 100µm resolution human brain dataset (#132)
# Description 

This is a redo of napari/napari.github.io#64. I have tried to keep the
original commit history. I think most comments there were addressed, but
it probably needs a new round of reviews to make sure it is current and
correct.

This is a tutorial on using Dask and Napari to view and process 100µm
resolution human brain dataset.

cc @DaliaAymanAmeen

# References
Redo of napari/napari.github.io#64


Co-authored-by: DaliaAymanAmeen <dalia.ayman.ameen@gmail.com>
Co-authored-by: Nicholas Sofroniew <sofroniewn@gmail.com>
Co-authored-by: Genevieve Buckley <30920819+GenevieveBuckley@users.noreply.github.com>
Co-authored-by: Talley Lambert <talley.lambert@gmail.com>
Co-authored-by: Dina Sh <1600540@eng.asu.edu.eg>
Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>