Add mp4 and png fallback to webm videos (#298)
# References and relevant issues

Closes #60 and https://github.com/napari/napari-sphinx-theme/issues/140

# Description

Adds mp4 and png fallback to all webm videos. Also moves them to the
appropriate folder for html builds. The downside here is that, in
addition to size (because we have more files now), the notebook view of
the guides can't seem to be able to play the videos (or use the fallback
png). I don't think this is something we use often (open the .md files
as notebooks) but ideally I'd like to keep it working.

EDIT: 
Videos have been updated as part of #125

- Since the fallback videos are now being created on deployment, the
only caveat left is the notebook observation above.