[Maint] fix build_and_deploy.yml paths (#384)
# References and relevant issues
closes: #383 

The deploy workflow "works" but is deploying nothing:

https://github.com/napari/docs/actions/runs/8563891050/job/23469798992#step:4:1833
> cp: no such file or directory: /home/runner/work/docs/docs/docs/*

The artifact is extracted to:

https://github.com/napari/docs/actions/runs/8563891050/job/23469798992#step:2:12
> Starting download of artifact to: /home/runner/work/docs/docs
> ...
> Extracting artifact entry:
/home/runner/work/docs/docs/_images/3D-button.png

There is confusing paths with docs and docs. It looks like the
deployment is relative to the last working dir and not the top-level.

# Description
In this PR I set the path of the download action to `html` (instead of
downloading by name, which shouldn't matter here) which should make
things clearer and then use that for the deployment.