[Maint] use lxml[html_clean] in requirements.txt (#397)
# References and relevant issues
Build & deploy is failing:

https://github.com/napari/docs/actions/runs/8608784588/job/23591705821?pr=390
This is because of the change to lxml. This was fixed in napari/napari,
but we ~~install from pypi~~ arn't using [dev] at the moment.


# Description
We have lxml in the requirements.txt so this PR just updates that to fix
the current issue, but we should revisit the actual
requirements/dependencies/install.