Replace use of properties with features (#425)
# References and relevant issues
Part of https://github.com/napari/napari/issues/3911

# Description
This replaces use of `properties`, `property_choices`, and
`current_properties` with `features` and `feature_defaults`. The
motivation for doing this now is to make the latest documentation
consistent with existing code and examples on `main` (here and on
napari/napari) for the upcoming v0.5 release. As we also plan to
deprecate properties in v0.5
(https://github.com/napari/napari/issues/3911), this documentation
update is necessary for that.