NAP-6: Updates after discussion and review (#312)
# Description
This PR updates NAP-6 after feedback on #77, discussion at the napari
hackathon (notes
[here](https://hackmd.io/fmKp0If5RkiwWIxYYRdKpg#Menu-contributions-notes)
but also linked in doc), and discussions in community meetings in recent
weeks.

## Still to be decided

- [x] `npe2` allows users to define new submenus as well as menu items.
**Should plugin developers be allowed to add new submenus to
contributable menus?** Decision: yes, but not to the top level `Layers`
menu, and the context menu shouldn't actually be contributable as it's
already very long.
- [x] It's not clear that the `Tracks` and `Classification` submenus are
as well defined as the rest. Classification at least I'm not even sure
how to describe. **Should we remove `Tracks` and `Classification` from
the initial list of contributable menus?** Decision: No, we like both!


---------

Co-authored-by: alisterburt <alisterburt@gmail.com>
Co-authored-by: Lucy Liu <jliu176@gmail.com>