Fix rendering example code (#441)
# References and relevant issues
Fixes #436

# Description
This fixes the rendering example code by adding a 2D layer rather than
replacing an existing 3D one. This ensures that the new layer is
properly set up (e.g. a corresponding a vispy layer is mapped to it).

This also improves the alt text for the related screenshots.