Update labels.md for polygon tool stopping behavior (#438)
# References and relevant issues
Updates docs for changes in:
https://github.com/napari/napari/pull/7019
https://github.com/napari/napari/pull/7023


# Description
In the two PR above, the polygon tool icon was changed from a triangle
to a polygon. Additionally, the double-click condition of the Labels
version of the tool was switched to match the Shapes one, with the
proximity feature being moved to Settings > Experimental.
I also updated the copy for the Shapes version of the tool, because I
noticed it doesn't mention double-click.