deploy docs on manual trigger (#462)
It turns out that we are building docs but not deploying docs when we merge
napari/napari PRs, and also on manually-triggered workflows. Only merges to
main on the napari/docs repo get deployed.

This PR also adds workflow_dispatch to the conditions in which built docs are
deployed. (This is also the mechanism by which merges in napari/napari trigger
docs building in the this repo.)