Add a release action
This action will build and deploy docs for a new release, copying
the built docs artifacts to both the dev and the {version} folders
in napari.github.io. It will also update the stable symlink to point
to the new version folder.

This action will be triggered by a new release tag pushed to the repo or
by manually trigerring the action with an input specifying the new version
(any input different from the default dev will trigger the action.)