Fix broken links detected by make linkcheck-files (#484)
# References and relevant issues
Related to
https://github.com/napari/docs/pull/474

# Description
Fixes a number of broken links detected by `make linkcheck-files`.

A few comments:
- it looks like this link does not exist anymore:
`further-resources/sample_data.md:17: [redirected permanently]
[digitalpathologyassociation.org/whole-slide-imaging-repository](https://digitalpathologyassociation.org/whole-slide-imaging-repository)
to
[https://digitalpathologyassociation.org?preview_error](https://digitalpathologyassociation.org/?preview_error)`.
I tried googling but only found this: http://www.wsirepository.org/ and
it doesn't load. Any ideas?
- How can we replace the following link?
`developers/architecture/app_model.md:288: [broken]
https://github.com/napari/napari/blob/main/napari/_app_model/_submenus.py:
404 Client Error: Not Found for url:
https://github.com/napari/napari/blob/main/napari/_app_model/_submenus.py`
I couldn't find SubmenuItem in the codebase, maybe I'm missing something
- https://github.com/potating-potato has probably deleted their github
account. I suggest we convert the links to static text if we want to
preserve history.
- These hackmd docs seem to only be readable if you are logged in:
https://hackmd.io/fmKp0If5RkiwWIxYYRdKpg and
https://hackmd.io/@talley/SJB_lObBi#What-is-a-virtual-environment Should
we make it readable by all, or ignore in the linkcheck?
- A few links will need to be fixed in npe2, I'll send a follow-up PR
for that.
- This file does not seem to exist anymore, from a search at the btrack
repo: `tutorials/tracking/cell_tracking.md:229: [broken]
https://github.com/quantumjot/btrack/blob/main/models/cell_config.json:
404 Client Error: Not Found for url:
https://github.com/quantumjot/btrack/blob/main/models/cell_config.json`.
Should we replace with a new link?
- I will add a regex to ignore redirects from docs paths to stuff like
"stable/" or "en/latest"

A few items are left, mostly links to old code. I will try to replace
with permalinks where I can, but not sure if I'll be able to find all of
them. If that's the case I might try to ping the original authors for an
appropriate substitution.

---------

Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>