Add linkcheck step to be run once a week (#474)
# References and relevant issues
Closes #439 
Closes #328
Depends on #484 

# Description
Adds a sphinx linkcheck step to be run once a week. The reported
failures will show up as issues in the repo.

This check can also be triggered manually. I am using the sphinx
linkchecker as it will check all links in the docs, including those
coming from docstrings.

You can see this in action on my fork, here: 
https://github.com/melissawm/napari-docs/issues/16