Add reference to architecture guide in the contributing guide (#537)
# References and relevant issues

In [napari issue
#7447](https://github.com/napari/napari/issues/7447#issuecomment-2543230977),
@psobolewskiPhD shared with me the architecture guide, saying not many
people know about it.

This adds a reference to the napari architecture guide to the main
Contributing guide index page, as a first-guess place a user might find
it. I myself did not know about this, in part because I'm a noob that
did not ever connect 'Architecture Guide' to something code related, and
therefore useful in my quest to start contributing. As such, having an
explicit mention of the guide and its purpose in different language
could be helpful.