Reorganize plugin landing page and remove redundant index file (#609)
# References and relevant issues
<!-- What relevant resources were used in the creation of this PR?
If this PR addresses an existing issue on the repo,
please link to that issue here as "Closes #(issue-number)".
If this PR adds docs for a napari PR please add a "Depends on <napari PR
link>" -->
Closes #595

# Description
1. Changes plugin landing page tabs to link immediatelly to
'finding_and_installing_plugins' and the napari hub, rather than to the
index of start_using_plugins, which contained nothing more than the same
information now on the landing page.
2. Removes index from start_using_plugins because it is not needed
3. Updates TOC to remove the index and using finding and installing
plugins as top level header

<!-- Final Checklist
- If images included: I have added [alt
text](https://webaim.org/techniques/alttext/)
If workflow, documentation build or deployment change:
- My PR is the minimum possible work for the desired functionality
- I have commented my code, to let others know what it does
-->