Move user-facing information to top of preferences and fix autogenerated UI images (#613)
# References and relevant issues
<!-- What relevant resources were used in the creation of this PR?
If this PR addresses an existing issue on the repo,
please link to that issue here as "Closes #(issue-number)".
If this PR adds docs for a napari PR please add a "Depends on <napari PR
link>" -->

# Description
<!-- What does this pull request (PR) do? Does it add new content,
improve/fix existing
context, improve/fix workflow/documentation build/deployment or
something else?
<!-- If relevant, please include a screenshot or a screen capture in
your content
change: "An image is worth a thousand words!" -->
<!-- You can use https://www.cockos.com/licecap/ or similar to create
animations. -->
<!-- You can also see a preview of the documentation changes you are
submitting by
clicking on "Details" to the right of the "Check the rendered docs
here!" check on your PR.-->

1. Moves user-facing information to the top of the autogenterated
preferences document.
2. Removes the redundant programmatic settings code block (unless we
also want to show the settings.SETTINGS version. I think as a learner,
understanding how `get_settings()` is used to set a singleton was more
important to my understanding)
3. Fixes/updates the autogenerated preferences dialog images to
correctly match the header to the image, and update the UI to correctly
show the selection in the list on the left.

<!-- Final Checklist
- If images included: I have added [alt
text](https://webaim.org/techniques/alttext/)
If workflow, documentation build or deployment change:
- My PR is the minimum possible work for the desired functionality
- I have commented my code, to let others know what it does
-->