Update conf.py to use packaging to detect dev vs release versions (#631)
# References and relevant issues
If you go to dev docs right now:
https://napari.org/dev/index.html
The sidebar link to the release notes is broken, I think because the
version is being treated as a release and not dev.
It points to:
https://napari.org/dev/release/release_0_6_0a0.html
which doesn't exist.
instead of the index.html which would be used for a dev release:
https://napari.org/dev/release/index.html

So in this PR I use packaging to check if a version is a dev version and
also uses packaging to get just the base version number of pre-releases,
so we will get release_0_6_0.html

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>