update left sidebar to persist resources (#636)
# Description

During the past few docs meetings and community meetings, we have been
talking about ways to provide users information about using napari and
engaging them in the napari community. One idea that kept coming up was
how to provide users easy access to troubleshooting information and
release notes.

After a recent docs meeting, @TimMonko came up with the idea to put a
few frequently used links in the left sidebar. He made a PR that would
add the links to the sidebar on the homepage. Someone else commented
that it would be nice to have the links on other pages.

I had played around with the napari sphinx theme and its templates a
couple of months ago. I had the idea to update the left sidebar's jinja
template to include a persistent section of frequently used links.
Here's a before and after view:

![Screenshot 2025-03-20 at 8 37
42 AM](https://github.com/user-attachments/assets/3b2b8568-9298-43af-bbe4-27f4badf0342)

![Screenshot 2025-03-20 at 8 46
27 AM](https://github.com/user-attachments/assets/65ac1e6f-0b17-4184-a862-ef2b8b574259)