Use env var to control if notebooks are executed and add docs, slim, slimfast, and slimgallery variants to Makefile (#646)
# References and relevant issues
closes: https://github.com/napari/docs/issues/625

# Description
In https://github.com/napari/docs/issues/625#issuecomment-2759683103
Juan noted that there is a flag in Myst NB to control whether notebooks
are executed.
This is a proof of concept using that.
Basically, the flag is set via ENV var. If it's not set, the flag is set
to `auto`, which is the default. See:
https://myst-nb.readthedocs.io/en/latest/configuration.html#execution

The effect is that now you can do something like:
`NB_EXECUTION_MODE="off" make html-noplot` and notebooks will not be
executed.

Additional slowness is from `prep_docs` which builds files for events,
preferences, and the UI arch diagrams. All of the source material is
from outside napari/docs.
As a result, I've added `--stubs` flag to all of the scripts, which will
generate a stub file if the .md file is missing. This ensures that there
are no warnings when trying to build the docs.

In order to give more control over building docs, I've added to the
Makefile:
- `prep-stubs` -- this runs `prep_docs.py --stubs`, to generate stubs
instead of the real files. Will not overwrite, so you can manually run
`prep-docs` if you want.
- `docs` -- this will do just napari/docs, but run notebooks (pops up
napari GUIs). Meaning: no gallery, prep_docs --stubs, nothing from
napari/napari basically
- `docs-live` -- same as above, but autobuild
- `slim` -- same as docs, but without notebook execution
- `slimfast` which is `slim` but with parallelization (`-j auto`). For
me this one runs in ~35 s.
- `slimfast-live` -- live version of `slimfast`. With `slimfast` being
fast, it runs an initial build. Great way to get started with editing
non-codecell content.
- tweaked `html-live` to be called `html-noplot-live`, reflecting its
actual role for consistency and to use `prep-stubs`, because those files
get ignored anyways. It *will* still execute notebooks during autobuild.

We have so many examples, that it's slow to build them all.
As a result, I modified `clean` to not remove them -- you can use
`clean-gallery` to do that.
Then I added:
- `slimgallery` -- will clean and then build the docs with the examples
gallery, but not run all the notebook cells in the rest of the docs. So
only the examples will pop the viewer.
- `slimgallery-live` will do the above, but as an autobuild. Will build
all examples that arn't built--the whole gallery. Then, only the example
you edit will be re-built when you edit it. Other pages will also
rebuild, but notebook cells will not run.
- `slimgallery-%` -- `%` is a capture, to a variable. This is a target
for `build-single-example` Here this will clean and build just a single
example, so you would do: `slimgallery-vortex` for the `vortex.py`
example.
- `slimgallery-live-%` -- this is the same as above, but will autobuild
the single example. Editing other examples will be ignored. Editing
other pages will work and trigger rebuild.

Finally, I updated the documentation and README for the new make
options.