Update roadmap index to highlight global roadmap  (#651)
# References and relevant issues
I always thought the roadmap page was just out of date and there was not
a roadmap since 0.4. Juan pointed out to me that its up to date on
Github.

# Description
This PR emphasizes the global roadmap by moving it as the most immediate
item. It also adds the For Contributors and For Funders part, which we
can opt to remove since that will require keeping both the roadmap
README and this up to date. This PR also tries to de-emphasize the
previous roadmaps.

Thank you Peter for slimfast-live. Made editing this such a breeze.


![image](https://github.com/user-attachments/assets/c921eb66-10d4-4599-a47e-0d79f0d6a541)

## Alternative

This removes for contributors and funders. I liked this less at first,
but maybe like it more now...

![Screenshot 2025-04-03
084657](https://github.com/user-attachments/assets/cad47363-39d2-4cc7-94f5-4e22cc76324d)