[bugfix] fix prep_docs logic to ensure npe2 files are written, but not over-written (#671)
# References and relevant issues
Fixes logic bug in https://github.com/napari/docs/pull/649, see
https://github.com/napari/docs/pull/649#discussion_r2041137415 that is
causing npe2 stubs to not be generated on CI runs, see:
https://github.com/napari/docs/actions/runs/14430288050/job/40464266128?pr=669#step:8:250

# Description
When testing #649 this I removed the `not` so that it would overwrite,
so I could iterate more easily.
I forgot to switch it back. make culpa.
But without the not, on a clean run, where the files don't exist, they
won't be generated -- this is the case of CI.
This PR puts the `not` back.