Add command palette to the Quick start guide and Viewer tutorial (#683)
# References and relevant issues
Closes: https://github.com/napari/docs/issues/569

# Description

In this PR I add a tip admonitions about the command palette and Juan's
great video to both the Quick start guide and the Viewer tutorial.
I think the location in the Quick start guide is pretty natural, but I'm
not sure about the Viewer tutorial. The viewer tutorial is rather
python/code cell centric. But I think the command palette is a pretty
nice way for a new user to discover things.