Update pan-zoom video with new UI, better data, and better framerate (#692)
The tutorial states "As `vispy` uses `OpenGL` and your graphics card,
panning and zooming are highly performant." But the corresponding video
is a resampled gif and looks choppy-af. This PR replaces that video with
one using a large dataset (that better shows off the zoom capabilities)
and 30fps recording, showing off the smoothness of the canvas.

The dataset comes from the NSF NOIRLab:

https://noirlab.edu/public/

Full data info:

```
url: https://noirlab.edu/public/images/noirlab2501a
title: DECam Deep View of the Antlia Cluster
description: |
    The Antlia Cluster (Abell S636) is a group of at least 230 galaxies located
    about 130 million light-years away in the direction of the constellation
    Antlia (the Air Pump). It hosts a rich variety of galaxy types, including
    lenticular galaxies, irregular galaxies and ultra-compact dwarfs. The
    cluster is dominated by two massive elliptical galaxies — NGC 3268 (center)
    and NGC 3258 (lower right).

    This image was taken with the 570-megapixel Department of Energy-fabricated
    Dark Energy Camera (DECam), mounted on the U.S. National Science Foundation
    (NSF) Víctor M. Blanco 4-meter Telescope at Cerro Tololo Inter-American
    Observatory in Chile, a Program of NSF NOIRLab. It captures only a portion
    of the 230 of galaxies that so far have been found to make up the Antlia
    Cluster.  DECam’s ultra-deep view showcases the variety of galaxies within
    and beyond the cluster in incredible detail.

    Credit:
    Dark Energy Survey/DOE/FNAL/DECam/CTIO/NOIRLab/NSF/AURA
    Image processing: R. Colombari & M. Zamani (NSF NOIRLab)
```

I do intend to make a napari plugin with these datasets but for now
let's just get this in. ;)