Fix comment and manual dispatch triggered build jobs (#723)
# References and relevant issues
Currently, comments are not properly triggering jobs.
The primary issue is that comment triggers are from the main branch
context and not PR context.
For circleCI it's even harder, because if you checkout the PR and then
run CircleCI it will still internal checkout the wrong branch.
Also reporting back to the PR was also not working.
I did some googling and sorted out a pathway using
[CircleCI-Public/trigger-circleci-pipeline-action
](https://github.com/CircleCI-Public/trigger-circleci-pipeline-action)
-- not sure how I missed it before?
Also consulted Claude a bit.

# Description
I ended up re-doing the whole action, so I gave it a new name. I think
the diff would be atrocious, so this is easier.
The main change here are as follows. First for the Circle part:
1. To use
https://github.com/CircleCI-Public/trigger-circleci-pipeline-action in
the action, we need to adjust the circleci/config.yaml to add the
pipeline parameters required by the action and also switch to using a
job parameter for the make target. Then we can use workflows for
circleCI that are conditional on the GH_Meta pipeline parameter. The net
result is that the right make target is used.
2. Use
https://github.com/CircleCI-Public/trigger-circleci-pipeline-action
which lets us pass the PR branch (this isn't well documented, but works)
via target-branch, as well as the desired make target in GH_Meta. The
best part is the status reporting then works for both the Circle job and
the redirector :tada:
We get the target-branch from the "PR details" step of the first job
`determine-make-target`

Importantly, this also sets us up for a followup PR to check for a label
and run full build. With this implementation that will be trivial by
comparison.

Now for the artifact build:
1. The key is the job is again getting those PR details. Then we use the
right branch for the reusable workflow and get the right artifact built.
For this we add the extra workflow_call parameter so we can pass the PR
detail to the job.
2. The issue is with the status reporting. Again the context being the
main branch of the repo, getting the reusable action status is tricky.
myrotvorets/set-commit-status-action ended up being very useful for this
simplifying a ton vs. the old approach.
Fun part:
- Added the eyes reaction to show that the comment was received.

I made the PR branch my main branch in my fork and then opened a PR vs
that, which you can find here:
https://github.com/psobolewskiPhD/napari-docs/pull/90
You can see that things came together and work!
Feel free to try a comment!

---------

Co-authored-by: Carol Willing <carolcode@willingconsulting.com>