[Enh] When `ready-to-merge` is applied, do a full build of docs (#745)
# References and relevant issues
Followup to https://github.com/napari/docs/pull/723 and
https://github.com/napari/docs/pull/669
This came up in a meeting and zulip, plus maybe a PR reviews?

# Description
Adds an action to check for ready-to-merge label and trigger a full
build when tagged.
This is simpler than the comment action because it should use the PR
context by default.
fingers crossed.
Note it's just a one-time full build. Pushes will still use the default,
slimfast.

I did testing in my fork:
https://github.com/psobolewskiPhD/napari-docs/pull/89
But as we saw before isn't guaranteed to work here! So we may need to
merge optimistically and then fix if it doesn't trigger correctly. 😬