Only build artifact on ready-to-merge label (#749)
# References and relevant issues
Alternative to https://github.com/napari/docs/pull/748

# Description
To avoid using the less secure pull_request_target trigger, we can just
avoid building the CircleCi preview.
The idea is that when reviewing, you would do the bot trigger to build
the needed preview. The ready-to-merge build would just be there to
verify that a full build doesn't fail.
Note: because a reusable action is used, it will be triggered from the
fork, but that should be ok as long as the fork branch is up to date
with napari/docs `main`, which is easy to ensure with the update button
in PRs.