Update napari.org homepage to remove the `imshow` "button" (#780)
# References and relevant issues
At SciPy 2025, Harry, a UI specialist, asked about the non-functional
button on the homepage. We were confused and realize he was referring to
the `imshow` command that is in a box that renders like a button:
<img width="641" height="347" alt="image"
src="https://github.com/user-attachments/assets/e896fb96-a243-4617-869a-d8e016abb187"
/>

The idea is that this is the code you can run to get the visual, but it
doesn't look like a code command (left justified with prompt) and the
code doesn't actually run either.

As a result, it was suggested to remove this and recover the
real-estate.

# Description

This PR removes the `figcaption` that adds a box with the `imshow`
command under the video animation.