Updates to NAP-9: Multiple Views (#730)
# References and relevant issues
Current (before this PR) rendered version of this NAP:
https://napari.org/stable/naps/9-multiple-canvases.html

You can click on the "Check the rendered docs here!" below to see the
rendered version of this update!

# Description
It's time to pick up this work again! A lot of the changes should
hopefully be self-explanatory, but there are a few general things to
note:

1. Naming change from "canvas" to "view": there were several discussions
about this, which I tried to summarize in the `Alternatives >
Terminology` section. Feel free to discuss and propose more!
2. There's now a lot more reliance on the Async work. When the NAP was
first drafted, we pinpointed the potential overlap and mentioned it, but
by now I think that work has advanced enough to see that we should
simply finish it and rely on it.
3. Switching from a single shared layerlist (as proposed by @aganders3
in his original draft in #249) to a per-view layerlist: this was my main
disagreement on that version of the implementation, and I tried to
highlight some of the reasons in the `Alternatives > Single LayerList`
section. I'd like particular attention/discussion on this part so we can
be more comprehensive on the pros/cons and reach a more consensus
decision.
4. I removed the architecture diagrams (sorry @aganders3 :sweat_smile:):
they were outdated, and I found they didn't help me much with following
the design logic due to the high level of detail and unrelated
methods/attributes being mentioned. If you disagree and think they
should be updated and added back in, let me know.

As usual for NAPs (as per [the NAP
workflow](https://napari.org/stable/naps/0-nap-process.html#nap-workflow)),
remember that this is just another iteration in the discussion, so this
PR will be merged relatively liberally, after enough discussion and
changes have accumulated. Declaring the NAP as `Accepted` is not part of
this PR.

---------

Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>